requirements.txt
aiogram==3.20.0
aiosqlite==0.20.0
ffmpeg-python==0.2.0
aiofiles==23.2.1
aiohttp==3.9.5
magic-filter==1.0.12
mutagen>=1.45.1
python-dotenv==1.0.0
Pillow==10.0.0